card can be saved

+++ ecommerce/src/main/java/com/example/ecommerce/dto/CardResponse.java
@@ -1,4 +1,4 @@
 package com.example.ecommerce.dto;
 
-public record CardResponse(Integer cardNo) {
+public record CardResponse(Long cardNo) {
 }
